src/ai_code_reviewer/logging_service/requirements.txt
# Logging Service Dependencies
# AI-Powered Python Code Reviewer - Asynchronous Logging Service

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0

# Message Queue
pika==1.3.2
redis==5.0.0  # Redis for persistent queue storage

# Database
psycopg2-binary==2.9.9

# Environment Configuration
python-dotenv==1.0.0

# Additional utilities
python-multipart==0.0.6

# Monitoring
prometheus-client==0.19.0

